# v-cloud-test-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vibbraneo",
        "contact": {
            "email": "diogorcazevedo@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://administracaodosistema.com.br/sv/api"
        },
        {
            "url": "http://127.0.0.1:8000/api/"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "vibbraneo@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/bid/getAllFromDeal/{id}": {
            "get": {
                "tags": [
                    "bid"
                ],
                "summary": "get all bids param Deal",
                "description": "List all bids from Deal",
                "operationId": "getAllBidsFromDeal",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bid/getAllFromUser/{id}": {
            "get": {
                "tags": [
                    "bid"
                ],
                "summary": "get all bids param User",
                "description": "List all bids from User",
                "operationId": "getAllBidsFromUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bid/getById/{id}": {
            "get": {
                "tags": [
                    "bid"
                ],
                "summary": "get bid param id",
                "description": "get bid by User",
                "operationId": "getBidById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id param ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bid/store/{deal_id}/{user_id}": {
            "post": {
                "tags": [
                    "bid"
                ],
                "summary": "store bid",
                "description": "create bid",
                "operationId": "bidStore",
                "parameters": [
                    {
                        "name": "deal_id",
                        "in": "path",
                        "description": "deal_id request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "id do usuário requerente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "store bid fields",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "value"
                                ],
                                "properties": {
                                    "accepted": {
                                        "type": "string",
                                        "format": "text",
                                        "enum": [
                                            "Venda",
                                            "Troca",
                                            "Desejo"
                                        ],
                                        "example": "Venda"
                                    },
                                    "value": {
                                        "type": "number",
                                        "format": "text",
                                        "example": "15000"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Proponho troca por iphone"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "descrition list"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/bid/accepted/{bid_id}": {
            "put": {
                "tags": [
                    "bid"
                ],
                "summary": "accepted bid update",
                "description": "accepted bid update",
                "operationId": "bidAccepted",
                "parameters": [
                    {
                        "name": "bid_id",
                        "in": "path",
                        "description": "bid_id request",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Pass accepted bid update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "accepted"
                                ],
                                "properties": {
                                    "accepted": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "accepted bid update"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}